% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co.R
\docType{data}
\name{co}
\alias{co}
\title{Tibble con i dati di co per 13 stazioni della regione LAZIO}
\format{
Un tibble con 8 colonne e 9503 osservazioni
}
\usage{
co
}
\description{
Tibble con i dati di co per 13 stazioni della regione LAZIO
}
\keyword{datasets}
